Started coding the MVP Design.
I am focusing on the TODAY page/tab. This is the page where users will be able to see all their tasks that need to be completed TODAY. I have the progress bar set up as well as the initial tabs and the main TODAY view.

There is still a lot more work to make it more beautiful. But the important stuff is functionality now. Then beautify. MVP.