Habit Flow Work.
Started working on the flow of adding a habit with a streak count. I need to make it display both on the Adventures Page and as a daily habit on the Today page. Still some issues with how the points accumulate.. And still need to make sure the daily habits and normal tasks are well distinguised.